requirements.txt
absl-py==1.4.0
altair==4.2.2
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
astunparse==1.6.3
attrs==22.1.0
autocommand==2.2.2
backcall==0.2.0
backports.csv==1.0.7
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.5
cachetools==5.3.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cheroot==9.0.0
CherryPy==18.8.0
click==8.1.3
colorama==0.4.5
contourpy==1.0.7
cryptography==39.0.1
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.2.1
docopt==0.6.2
email-validator==1.3.0
entrypoints==0.4
executing==1.1.1
fastapi==0.85.0
fastjsonschema==2.16.2
feedparser==6.0.10
filelock==3.8.0
flatbuffers==23.1.21
fonttools==4.38.0
future==0.18.3
fuzzy-c-means==1.6.4
fuzzywuzzy==0.18.0
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.16.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h11==0.14.0
h5py==3.8.0
httptools==0.5.0
huggingface-hub==0.10.1
idna==3.4
importlib-metadata==6.0.0
inflect==6.0.2
ipykernel==6.16.0
ipython==8.5.0
ipython-genutils==0.2.0
jaraco.classes==3.2.3
jaraco.collections==3.8.0
jaraco.context==4.3.0
jaraco.functools==3.6.0
jaraco.text==3.11.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.10
jsonschema==4.16.0
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-pygments==0.2.2
keras==2.11.0
kiwisolver==1.4.4
libclang==15.0.6.1
lxml==4.9.2
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.1
matplotlib==3.7.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.4
more-itertools==9.0.0
mysqlclient==2.1.1
nbclient==0.7.0
nbconvert==7.2.1
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==2.8.7
nltk==3.8.1
notebook==6.4.12
npm==0.1.1
numpy==1.23.3
oauthlib==3.2.2
opencv-python==4.7.0.68
opt-einsum==3.3.0
optional-django==0.1.0
packaging==21.3
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
Pattern==3.6
pdfminer.six==20221105
pickle-mixin==1.0.2
pickleshare==0.7.5
Pillow==9.4.0
pipreqs==0.4.11
portend==3.1.0
prometheus-client==0.14.1
prompt-toolkit==3.0.31
protobuf==3.19.6
psutil==5.9.2
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.2
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
pywin32==304
pywinpty==2.0.8
PyYAML==6.0
pyzmq==24.0.1
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rich==13.3.1
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.3
semver==2.13.0
Send2Trash==1.8.0
sgmllib3k==1.0.0
six==1.16.0
sklearn==0.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.5.1
starlette==0.20.4
streamlit==1.19.0
tabulate==0.8.10
tempora==5.2.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
terminado==0.16.0
threadpoolctl==3.1.0
tinycss2==1.1.1
tokenizers==0.13.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
transformers==4.24.0
typer==0.4.2
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.12
uvicorn==0.18.3
validators==0.20.0
watchdog==2.2.1
watchfiles==0.17.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
websockets==10.3
Werkzeug==2.2.3
wrapt==1.14.1
yarg==0.1.9
zc.lockfile==3.0
zipp==3.14.0
